CDK-514: Update dataset example.
This updates the example to use Datasets and views. It also changes
logging so that the Avro file reader's debug messages can be used to
demonstrate that only one partition is read in the partitioning section.

Partitioning is now demonstrated with favoriteColor rather than by a
hash partitioner. This is easier to understand and demonstrates the
storage layout more clearly.