CDK-534: Switch demo crunch jobs to use views.
This implements the same command-line arguments as the previous version,
but the getLatestPartition method is replaced with viewFromUri. The new
version relativizes the URI and then uses the path's key/value pairs to
create a view.

The 'latest' partition is now time-based. Before, the newest partition
in the data was always used, but this wasn't correct and led to
duplicate sessions when the oozie job ran with no data. Because of this,
the demo-crunch tool differs slightly from the demo-oozie tool. The
demo-crunch tool will process everything before the current minute
without a lower bound. That ensures that running with no arguments
always processes some data rather than aborting because there was no
traffic in the last minute.

This depends on kite-data-core fixes for CDK-532 and CDK-536.